cyrusdb: Add cyrusdb benchmarks.
This commit adds a basic framework and an intial set of benchmarks for
write. Currently supports benchmarking either on `twoskip` or
`zeroskip`. There's no particular reason that the benchmark shouldn't
run on other backends, this is just an artificial limit. We will add
support for all the backends at some point.

The benchmarking tool is based on the benchmarking tool in LevelDB.

Currently the benchmarks that are run are limited to benchmarking
writes, and has the following benchmarks:
 * writeseq       - write values in sequential key order
 * writeseqtxn    - write values in sequential key order in separate transactions
 * writerandom    - write values in random key order
 * writerandomtxn - write values in random key order in separate transactions
 * write100k      - write values 100K long in random key order